src/requirements.txt
synapseclient>=3.0.0
google-cloud-secret-manager>=2.16.0
google-cloud-storage>=2.10.0
google-api-python-client>=2.100.0
